hessian unshared optimization
git-svn-id: svn://svn.caucho.com/resin/trunk@8507 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/server/http/AbstractHttpRequest.java
@@ -702,7 +702,7 @@ public abstract class AbstractHttpRequest
 
   protected void setContentLength(CharSegment value)
   {
-    int contentLength = 0;
+    long contentLength = 0;
     int ch;
     int i = 0;
